rename modules related to sensor management actions
