+++ stonesoup/sensor/action/dwell_action.py
@@ -4,7 +4,7 @@ from typing import Iterator
 
 import numpy as np
 
-from ...sensormanager.actionable import Action, RealNumberActionGenerator
+from ...sensormanager.action import Action, RealNumberActionGenerator
 from ...base import Property
 from ...functions import mod_bearing
 from ...types.angle import Angle, Bearing
