+++ stonesoup/sensor/action/tests/test_dwell_action.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from ..dwell_action import DwellActionsGenerator, ChangeDwellAction
-from ....sensormanager.actionable import Actionable, ActionableProperty
+from ....sensormanager.action import Actionable, ActionableProperty
 from ...base import Property
 from ....types.angle import Bearing, Angle
 from ....types.array import StateVector
